refactor: Bulk of logic to command
